requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
sqlmodel==0.0.22
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.3.6
alembic==1.13.3

# Google Gemini
google-generativeai==0.8.3

# WhatsApp Client
httpx==0.27.2
aiohttp==3.10.10

# Scheduling
apscheduler==3.10.4
pytz==2024.2

# Utilities
tenacity==9.0.0
structlog==24.4.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.3
